update client specification tests
